Move download stats to a plugin, add support for monthly-rounded stats
